pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.10",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
